requirements.txt
streamlit
groq
duckduckgo-search
arxiv
wikipedia
python-dotenv